// src/interfaces/teams.ts
export interface ITeam {
  idTeam: string;
  idSoccerXML: string;
  idAPIfootball: string;
  intLoved: string;
  strTeam: string;
  strTeamShort: string;
  strAlternate: string;
  intFormedYear: string;
  strSport: string;
  strLeague: string;
  idLeague: string;
  strLeague2: string;
  idLeague2: string;
  strLeague3: string;
  idLeague3: string;
  strLeague4: string;
  idLeague4: string;
  strLeague5: string;
  idLeague5: string;
  strLeague6: string;
  idLeague6: any;
  strLeague7: string;
  idLeague7: any;
  strDivision: any;
  idVenue: string;
  strStadium: string;
  strKeywords: string;
  strRSS: string;
  strLocation: string;
  intStadiumCapacity: string;
  strWebsite: string;
  strFacebook: string;
  strTwitter: string;
  strInstagram: string;
  strDescriptionEN: string;
  strDescriptionDE: string;
  strDescriptionFR: string;
  strDescriptionCN: any;
  strDescriptionIT: string;
  strDescriptionJP: string;
  strDescriptionRU: string;
  strDescriptionES: string;
  strDescriptionPT: string;
  strDescriptionSE: any;
  strDescriptionNL: any;
  strDescriptionHU: any;
  strDescriptionNO: string;
  strDescriptionIL: any;
  strDescriptionPL: any;
  strKitColour1: string;
  strKitColour2: string;
  strKitColour3: string;
  strGender: string;
  strCountry: string;
  strBadge: string;
  strLogo: string;
  strFanart1: string;
  strFanart2: string;
  strFanart3: string;
  strFanart4: string;
  strBanner: string;
  strKit: string;
  strYoutube: string;
  strLocked: string;
}